Fixing various runtime exceptions that could occur after invoking !bb
it could happen that the server already had closed the socket connection
but the server's listeners were still active. in that case useres could
still be fetched due to a caught event which resulted in a runtime
error. the reason for this exception is because of the closed connection
and the circumstance, that the User.all method uses a api method that
required an acitive server connetion.

i fixed this issue by introducing a nil_user that is returned whenever
no user could be fetched and making access to the bot's instance
variable @is_started mutex.