Allow to define custom permission lvls defined by the ts3 server
if existing, make use of the server_groups#sort_id value otherwise use
the group id as the sorting parameter.

this change makes it possible to have a server specified permission
system. before this change pigeon was considering the ids of the default
groups as the indicator of the permission lvl - which is true for a
default ts3 server configuration. however, when defining custom
permissions, this assumption is no more valid.

info: permission levels are defined as the follows:

a permission is basically a integer value. the lowever its value,
the higher the permission level (like in sport rankings).
a client may invoke/perform certain pigeon actions if and only if his
permission lvl is at least as high as the required permission level.