Implemented server functionality allowing to listen to external messages
Implemented a socket server that is listening to incoming messages from external
sources on a specified port. currently the port is hard-coded and the
bot processes every incoming message as it is and sends them in the
current channel. addidionally, it accepts messages comming from every
external source. introducing a whitelist would make sense here.